Rename include/tr_hub_parser to include/teraranger_hub
