+++ package.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0"?>
 <package>
-  <name>tr_hub_parser</name>
+  <name>teraranger_hub</name>
   <version>1.0.0</version>
-  <description>The tr_hub_parser package</description>
+  <description>The teraranger_hub package</description>
 
   <maintainer email="pierre-louis.kabaradjian@terabee.com">Pierre-Louis Kabaradjian</maintainer>
   <maintainer email="krzysztof.zurad@terabee.com">Krzysztof Zurad</maintainer>
@@ -24,7 +24,7 @@
   <build_depend>rospy</build_depend>
 
   <test_depend>rosunit</test_depend>
-  
+
   <run_depend>message_runtime</run_depend>
   <run_depend>roscpp</run_depend>
   <run_depend>sensor_msgs</run_depend>
